Move plugshare code to a module for easier notebook testing and streamline code
